fix: fix pr comment
