+++ packages/client/src/assets/styles/buttons.tsx
@@ -1,4 +1,4 @@
-export const loadMoreBtnSx = {
+export const loadMoreBtnStyles = {
   fontSize: '16px',
   fontWeight: 700,
   background: 'none',
